Allow independent viewport and light regions in C4FoWRegion (#1211)
With this change, an additional rectangle is stored in C4FoWRegion that
represents the area covered by the viewport in fractional floating point
coordinates. This allows the light texture to be created for an arbitrary
portion of the landscape, and the coordinate transformations for the
shaders will still work.

Also, since the additional rectangle uses floating point precision, the
computed coordinate transformations do now give the exact same result as for
the landscape pixel-by-pixel, and there should not be any offsets left.

I also hope that this change improves or fixes the single-pixel-lines of sky
that are sometimes seen at the edges of the viewport.

+++ src/game/C4Viewport.h
@@ -31,7 +31,7 @@ public:
 	// "display" coordinates
 	int32_t ViewWdt,ViewHgt;
 	// position of landscape border (left,top,right, bottom) in viewport. 0 if there is not border
-	int32_t BorderLeft, BorderTop, BorderRight, BorderBottom;
+	float BorderLeft, BorderTop, BorderRight, BorderBottom;
 	int32_t DrawX,DrawY;
 	// facets used for last drawing operations
 	C4TargetFacet last_game_draw_cgo, last_gui_draw_cgo;
